refactor: 테스트 수정 및 도메인 디렉토리 구조 변경
