+++ src/main/java/com/backend/detailgoal/domain/event/AlarmEvent.java
@@ -1,4 +1,4 @@
-package com.backend.detailgoal.domain;
+package com.backend.detailgoal.domain.event;
 
 public record AlarmEvent(
         String uid,
