+++ src/main/java/com/backend/global/event/ReminderEventHandler.java
@@ -1,7 +1,6 @@
 package com.backend.global.event;
 
-import com.backend.detailgoal.domain.AlarmEvent;
-import com.backend.goal.domain.ReminderEvent;
+import com.backend.goal.domain.event.ReminderEvent;
 import com.backend.infrastructure.fcm.FcmService;
 import com.backend.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
