+++ src/main/java/com/backend/global/scheduler/SchedulerService.java
@@ -1,11 +1,10 @@
 package com.backend.global.scheduler;
 
 import com.backend.detailgoal.application.dto.response.DetailGoalAlarmResponse;
-import com.backend.detailgoal.domain.AlarmEvent;
-import com.backend.detailgoal.domain.DetailGoalQueryRepository;
+import com.backend.detailgoal.domain.event.AlarmEvent;
+import com.backend.detailgoal.domain.repository.DetailGoalQueryRepository;
 import com.backend.goal.domain.Goal;
-import com.backend.goal.domain.GoalQueryRepository;
-import com.backend.goal.domain.ReminderEvent;
+import com.backend.goal.domain.repository.GoalQueryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -17,8 +16,6 @@ import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import java.util.Objects;
-import java.util.Random;
 
 @Slf4j
 @Service
