+++ src/main/java/com/backend/goal/application/GoalService.java
@@ -6,6 +6,11 @@ import com.backend.goal.application.dto.response.GoalListResponse;
 import com.backend.goal.application.dto.response.RetrospectEnabledGoalCountResponse;
 import com.backend.goal.domain.*;
 import com.backend.goal.application.dto.response.GoalResponse;
+import com.backend.goal.domain.enums.GoalStatus;
+import com.backend.goal.domain.event.RemoveRelatedDetailGoalEvent;
+import com.backend.goal.domain.repository.GoalListResponseDto;
+import com.backend.goal.domain.repository.GoalQueryRepository;
+import com.backend.goal.domain.repository.GoalRepository;
 import com.backend.goal.presentation.dto.GoalRecoverRequest;
 import com.backend.goal.presentation.dto.GoalSaveRequest;
 import com.backend.goal.presentation.dto.GoalUpdateRequest;
