+++ src/main/java/com/backend/goal/application/dto/response/GoalCountResponse.java
@@ -1,11 +1,8 @@
 package com.backend.goal.application.dto.response;
 
-import com.backend.goal.domain.GoalStatus;
-import io.swagger.v3.oas.annotations.StringToClassMapItem;
+import com.backend.goal.domain.enums.GoalStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.media.SchemaProperty;
 
-import java.util.HashMap;
 import java.util.Map;
 
 public record GoalCountResponse(
