+++ src/main/java/com/backend/goal/domain/enums/GoalStatus.java
@@ -1,4 +1,4 @@
-package com.backend.goal.domain;
+package com.backend.goal.domain.enums;
 
 public enum GoalStatus {
