+++ src/main/java/com/backend/goal/domain/event/ReminderEvent.java
@@ -1,4 +1,4 @@
-package com.backend.goal.domain;
+package com.backend.goal.domain.event;
 
 public record ReminderEvent(
