--- src/main/resources/application-prod.yml
@@ -1,34 +1,0 @@
-logging-module:
-  version: 0.0.1
-
-springdoc:
-  swagger-ui:
-    groups-order: DESC
-    disable-swagger-default-url: true
-    display-request-duration: true
-    operations-sorter: method
-  api-docs:
-    path: /api-docs
-  show-actuator: true
-  default-consumes-media-type: application/json;charset=UTF-8
-  default-produces-media-type: application/json;charset=UTF-8
-
-jwt:
-  secret: dGhpcyBpcyBhIHNhbXBsZSAyNTYtYml0IEJhc2U2dGhpcyBpcyBhIHNhbXBsZSAyNTYtYml0IEJhc2U2
-
-spring:
-  jpa:
-    hibernate:
-      ddl-auto: create
-  datasource:
-    driver-class-name: org.mariadb.jdbc.Driver
-    url: ${datasource}
-    username: ${username}
-    password: ${password}
-  data:
-    redis:
-      host: localhost
-      port: 6379
-
-fcm:
-  certification: firebase/fcm_key.json
