+++ src/test/java/com/backend/detailgoal/application/DetailGoalServiceTest.java
@@ -4,13 +4,13 @@ import com.backend.detailgoal.application.dto.response.DetailGoalListResponse;
 import com.backend.detailgoal.application.dto.response.DetailGoalResponse;
 import com.backend.detailgoal.application.dto.response.GoalCompletedResponse;
 import com.backend.detailgoal.domain.DetailGoal;
-import com.backend.detailgoal.domain.DetailGoalRepository;
+import com.backend.detailgoal.domain.repository.DetailGoalRepository;
 import com.backend.detailgoal.presentation.dto.request.DetailGoalSaveRequest;
 import com.backend.detailgoal.presentation.dto.request.DetailGoalUpdateRequest;
 import com.backend.global.DatabaseCleaner;
 import com.backend.goal.domain.Goal;
-import com.backend.goal.domain.GoalRepository;
-import com.backend.goal.domain.GoalStatus;
+import com.backend.goal.domain.repository.GoalRepository;
+import com.backend.goal.domain.enums.GoalStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
