+++ src/test/java/com/backend/scheduler/SchedulerServiceTest.java
@@ -3,11 +3,8 @@ package com.backend.scheduler;
 import com.backend.global.scheduler.SchedulerService;
 import com.backend.goal.application.GoalService;
 import com.backend.goal.domain.Goal;
-import com.backend.goal.domain.GoalQueryRepository;
-import com.backend.goal.domain.GoalRepository;
-import com.backend.goal.domain.GoalStatus;
-import com.backend.goal.presentation.dto.GoalSaveRequest;
-import org.assertj.core.api.Assertions;
+import com.backend.goal.domain.repository.GoalRepository;
+import com.backend.goal.domain.enums.GoalStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
